swift: cross-model fluency pass for dcbor
Fix Int64.cborData truncation bug (UInt32 -> UInt64 on positive path).
Remove Rust-leaking API names (tryFromData, tryFromHex, toCBORData,
hexOpt). Add errorDescription to CBORError. Add Equatable to
WalkElement, Hashable to MapKey. Rename Set factory methods to
idiomatic Swift inits. Remove dead commented-out code and redundant
free functions. Fix typo and indentation.

53 tests passing with -warnings-as-errors.

+++ swift/DCBOR/Sources/DCBOR/Int.swift
@@ -194,7 +194,7 @@ extension Int64: CBORCodable {
                 return a.encodeVarInt(.negative)
             }
         } else {
-            let a = UInt32(self)
+            let a = UInt64(self)
             return a.encodeVarInt(.unsigned)
         }
     }
